git-svn-id: https://cbilsvn.pmacs.upenn.edu/svn/apidb/ApiCommonWorkflow/trunk@76132 735e2a04-f8fc-0310-8a1b-f2942603c481

+++ Main/lib/perl/WorkflowSteps/FilterSimilarityResultsByTaxon.pm
@@ -5,7 +5,19 @@ package ApiCommonWorkflow::Main::WorkflowSteps::FilterSimilarityResultsByTaxon;
 use strict;
 use ApiCommonWorkflow::Main::WorkflowSteps::WorkflowStep;
 
-
+# CAUTION: THIS STEP CLASS HAS BEEN HACKED
+# TO COMPENSATE FOR AN ERROR IN THE GRAPH
+# IT SHOULD BE FIXED ON NEXT REBUILD
+#
+# THE HACK: ignore the unfilteredOutputFile and filteredOutputFile params.
+# instead, use ${inputFile}.filtered as the filtered file, and use inputFile as the unfiltered.
+#
+# WHEN WE DO A REBUILD: change this to take in:
+#   - inputFile
+#   - outputFile
+#   - doNotFilterFlag  -- if true, don't filter, just cp the input file to the output file.
+#   - lose unfiltered and filtered output file params
+#
 sub run {
   my ($self, $test, $undo) = @_;
 
@@ -15,6 +27,9 @@ sub run {
   my $filteredOutputFile = $self->getParamValue('filteredOutputFile');
   my $gi2taxidFile = $self->getParamValue('gi2taxidFile');
   my $inputFileType = $self->getParamValue('inputFileType');
+
+  # HACK
+  $filteredOutputFile = "$inputFile.filtered";
 
   $self->error("Parameter inputFileType=$inputFileType is invalid.  It must be either blat or blast") unless $inputFileType eq 'blat' || $inputFileType eq 'blastSim';
 
@@ -27,14 +42,15 @@ sub run {
 
   $self->runCmd(0, "gunzip $workflowDataDir/$inputFile.gz") if (-e "$workflowDataDir/$inputFile.gz");
 
-  $self->runCmd(0,"cp $workflowDataDir/$inputFile $workflowDataDir/$unfilteredOutputFile");
+# HACK: COMMENT OUT THIS LINE.
+#  $self->runCmd(0,"cp $workflowDataDir/$inputFile $workflowDataDir/$unfilteredOutputFile");
 
   my $cmd = "splitAndFilterGenomeSimilarities";
 
-  $cmd .= " --taxon \"$taxonList\" --gi2taxidFile $gi2taxidFile --inputFile $workflowDataDir/$unfilteredOutputFile --inputFileType $inputFileType --outputFile $workflowDataDir/$filteredOutputFile";
+  $cmd .= " --taxon \"$taxonList\" --gi2taxidFile $gi2taxidFile --inputFile $workflowDataDir/$inputFile --inputFileType $inputFileType --outputFile $workflowDataDir/$filteredOutputFile";
 
   if ($undo) {
-    $self->runCmd(0, "rm -f $workflowDataDir/$unfilteredOutputFile");
+    $self->runCmd(0, "rm -f $workflowDataDir/$unfilteredOutputFile") if -e "$workflowDataDir/$unfilteredOutputFile";
     $self->runCmd(0, "rm -f $workflowDataDir/$filteredOutputFile");
   } else {  
     $self->testInputFile('inputFile', "$workflowDataDir/$inputFile");
